+++ src/mantle/testing/class-utils.php
@@ -438,6 +438,19 @@ class Utils {
 	}
 
 	/**
+	 * Check if we're running in a CI (Continuous Integration) environment.
+	 */
+	public static function is_ci(): bool {
+		return (
+			! empty( $_SERVER['GITHUB_ENV'] )
+			|| ( ! empty( $_SERVER['CI'] ) && in_array( $_SERVER['CI'], [ 'true', '1' ], true ) )
+			|| ! empty( $_SERVER['GITHUB_REPOSITORY_OWNER'] )
+			|| ! empty( $_SERVER['GITHUB_WORKFLOW'] )
+			|| ! empty( $_SERVER['GITHUB_EVENT_NAME'] )
+		);
+	}
+
+	/**
 	 * Run a system command and return the output.
 	 *
 	 * @param string|string[] $command Command to run.